projectBuiding楼栋层的 cud service接口及context实现, 把之前我写的service方法全改为void 顺便修改了一下注释.

+++ src/main/java/com/liefeng/property/domain/project/ProjectContext.java
@@ -91,12 +91,15 @@ public class ProjectContext {
 	}
 
 	public ProjectVo update() {
-		ProjectPo projectPo = MyBeanUtil.createBean(project, ProjectPo.class);
-		projectRepository.save(projectPo);
+		if(project != null){
+			ProjectPo projectPo = MyBeanUtil.createBean(project, ProjectPo.class);
+			projectRepository.save(projectPo);
+		}
 		return project;
 	}
 
 	public void delete() {
-		projectRepository.delete(projectId);
+		if(projectId != null)
+            projectRepository.delete(projectId);
 	}
 }
